Load secrets only if a secrets store is configured

+++ sensor/run.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
 # Load information from secrets plugins
-python3 ./src/autoconfigure.py secrets
-sleep 3
-daprd --log-level debug --components-path /app/components/secrets --app-id $1 &
-DAPR_PID=$!
-sleep 3
-`python3 ./src/getSecrets.py` # environment variables are applied from stdout
-kill -SIGTERM "$DAPR_PID"
-sleep 5
+if python3 ./src/autoconfigure.py secrets
+then
+    sleep 3
+    daprd --log-level debug --components-path /app/components/secrets --app-id $1 &
+    DAPR_PID=$!
+    sleep 3
+    `python3 ./src/getSecrets.py` # environment variables are applied from stdout
+    kill -SIGTERM "$DAPR_PID"
+    sleep 5
+fi
 
 # Initialize output plugins
 python3 ./src/autoconfigure.py output
